refactor: package composer testing module (#5309)
* add testing utilities

* make setup.py consistent with requriements-test

* try moving package to top level

* remove egg info from commit

* add py_modules field to setup.py

* remove __init__.py from dagtestutils

* rename unit_testing

* modify setup.py

* bring back unit_testing name

* move under composer directory

* update subdirectory in requirements and setup

* reneame to dag_test_utils

* add readme, fix lint

* rename module (tests should fail)

* complete rename to internal_unit_testing

* fix lint oops

* remove branch reference
