+++ composer/dag_test_utils/setup.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2021 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,15 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import internal_unit_testing
+from setuptools import find_packages
+from setuptools import setup
 
-
-def test_dag_import():
-    """Test that the DAG file can be successfully imported.
-
-    This tests that the DAG can be parsed, but does not run it in an Airflow
-    environment. This is a recommended confidence check by the official Airflow
-    docs: https://airflow.incubator.apache.org/tutorial.html#testing
-    """
-    from . import simple as module
-    internal_unit_testing.assert_has_valid_dag(module)
+setup(
+    name="dag_test_utils",
+    version="0.0.1",
+    url="git@github.com:GoogleCloudPlatform/python-docs-samples.git#egg=dag_test_utils&subdirectory=composer/dag_test_utils",
+    author="Google Cloud Platform",
+    description="Utility used to unit test example Apache Airflow DAGs",
+    packages=find_packages(),
+    py_modules=['internal_unit_testing'],
+    install_requires=['apache-airflow[gcp]']
+)
