+++ composer/workflows/dataproc_workflow_template_instantiate_operator_tutorial_test.py
@@ -14,7 +14,7 @@
 
 from airflow import models
 
-from . import unit_testing
+import internal_unit_testing
 
 
 def test_dag_import():
@@ -27,4 +27,4 @@ def test_dag_import():
     models.Variable.set("project_id", "example-project")
     from . import dataproc_workflow_template_instantiate_operator_tutorial as module
 
-    unit_testing.assert_has_valid_dag(module)
+    internal_unit_testing.assert_has_valid_dag(module)
