+++ composer/workflows/use_local_deps_test.py
@@ -15,9 +15,8 @@
 import os.path
 import sys
 
+import internal_unit_testing
 import pytest
-
-from . import unit_testing
 
 
 @pytest.fixture(scope='module', autouse=True)
@@ -40,4 +39,4 @@ def test_dag_import():
     docs: https://airflow.incubator.apache.org/tutorial.html#testing
     """
     from . import use_local_deps as module
-    unit_testing.assert_has_valid_dag(module)
+    internal_unit_testing.assert_has_valid_dag(module)
